Add NavMeshAreaIndex enum
In addition to NavMeshAreas as a mask of areas, provide NavMeshAreaIndex
as the index for each area. This value is useful to assign to several
elements in NavMeshComponents (NavMeshSurface, NavMeshLink,
NavMeshModifier) -- they use the index instead of a mask.

The use of Split() is not ideal. For better perf, we could overwrite
the single line inside the enum instead of writing the whole thing out
or output an array of lines instead of a single string. Not sure which
is ideal, but I don't think either are necessary right now (this
processing only occurs after changes to the area list).